Enum rcgen::RcgenError

pub enum RcgenError {
    CouldNotParseCertificate,
    CouldNotParseKeyPair,
    KeyGenerationUnavailable,
    UnsupportedSignatureAlgorithm,
    RingUnspecified,
    CertificateKeyPairMismatch,
    Time,
    PemError(PemError),
    // some variants omitted
}

The error type of the rcgen crate

Variants

CouldNotParseCertificate

The given certificate couldn’t be parsed

CouldNotParseKeyPair

The given key pair couldn’t be parsed

KeyGenerationUnavailable

There is no support for generating keys for the given algorithm

UnsupportedSignatureAlgorithm

The requested signature algorithm is not supported

RingUnspecified

Unspecified ring error

CertificateKeyPairMismatch

The provided certificate’s signature algorithm is incompatible with the given key pair

Time

Time conversion related errors

PemError(PemError)

Error from the pem crate