rcgen

Struct CertificateRevocationListParams

pub struct CertificateRevocationListParams {
    pub this_update: OffsetDateTime,
    pub next_update: OffsetDateTime,
    pub crl_number: SerialNumber,
    pub issuing_distribution_point: Option<CrlIssuingDistributionPoint>,
    pub revoked_certs: Vec<RevokedCertParams>,
    pub key_identifier_method: KeyIdMethod,
}

Parameters used for certificate revocation list (CRL) generation

Fields
this_update: OffsetDateTime

Issue date of the CRL.

next_update: OffsetDateTime

The date by which the next CRL will be issued.

crl_number: SerialNumber

A monotonically increasing sequence number for a given CRL scope and issuer.

issuing_distribution_point: Option<CrlIssuingDistributionPoint>

An optional CRL extension identifying the CRL distribution point and scope for a particular CRL as described in RFC 5280 Section 5.2.51.

https://datatracker.ietf.org/doc/html/rfc5280#section-5.2.5 ↩

revoked_certs: Vec<RevokedCertParams>

A list of zero or more parameters describing revoked certificates included in the CRL.

key_identifier_method: KeyIdMethod

Method to generate key identifiers from public keys

Defaults to SHA-256.

Implementations

impl CertificateRevocationListParams

pub fn signed_by( self, issuer: &Certificate, issuer_key: &KeyPair, ) -> Result<CertificateRevocationList, Error>

Serializes the certificate revocation list (CRL).

Including a signature from the issuing certificate authority’s key.