rcgen

Struct CertificateSigningRequestParams

pub struct CertificateSigningRequestParams {
    pub params: CertificateParams,
    pub public_key: PublicKey,
}

Parameters for a certificate signing request

Fields
params: CertificateParams

Parameters for the certificate to be signed.

public_key: PublicKey

Public key to include in the certificate signing request.

Implementations

impl CertificateSigningRequestParams

pub fn from_pem(pem_str: &str) -> Result<Self, Error>

Parse a certificate signing request from the ASCII PEM format

See from_der for more details.

pub fn from_der(csr: &CertificateSigningRequestDer<'_>) -> Result<Self, Error>

Parse a certificate signing request from DER-encoded bytes

Currently, this only supports the Subject Alternative Name extension. On encountering other extensions, this function will return an error.

rustls_pemfile::csr() is often used to obtain a CertificateSigningRequestDer from PEM input. If you already have a byte slice containing DER, it can trivially be converted into CertificateSigningRequestDer using the Into trait.

pub fn signed_by( self, issuer: &Certificate, issuer_key: &KeyPair, ) -> Result<Certificate, Error>

Generate a new certificate based on the requested parameters, signed by the provided issuer.

The returned certificate will have its issuer field set to the subject of the provided issuer, and the authority key identifier extension will be populated using the subject public key of issuer. It will be signed by issuer_key.

Note that no validation of the issuer certificate is performed. Rcgen will not require the certificate to be a CA certificate, or have key usage extensions that allow signing.

The returned Certificate may be serialized using Certificate::der and Certificate::pem.