rcgen

Type Alias RcgenError

pub type RcgenError = Error;

👎Deprecated: Renamed to Error. We recommend to refer to it by fully-qualifying the crate: rcgen::Error.

Type-alias for the old name of Error.

Aliased Type

pub enum RcgenError {

    CouldNotParseCertificate,
    CouldNotParseCertificationRequest,
    CouldNotParseKeyPair,
    InvalidNameType,
    InvalidAsn1String(InvalidAsn1String),
    InvalidIpAddressOctetLength(usize),
    KeyGenerationUnavailable,
    UnsupportedExtension,
    UnsupportedSignatureAlgorithm,
    RingUnspecified,
    RingKeyRejected(String),
    Time,
    PemError(String),
    RemoteKeyError,
    UnsupportedInCsr,
    InvalidCrlNextUpdate,
    IssuerNotCrlSigner,
    X509(String),

}

Variants

CouldNotParseCertificate

The given certificate couldn’t be parsed

CouldNotParseCertificationRequest

The given certificate signing request couldn’t be parsed

CouldNotParseKeyPair

The given key pair couldn’t be parsed

InvalidNameType

Invalid subject alternative name type

InvalidAsn1String(InvalidAsn1String)

Invalid ASN.1 string

InvalidIpAddressOctetLength(usize)

An IP address was provided as a byte array, but the byte array was an invalid length.

KeyGenerationUnavailable

There is no support for generating keys for the given algorithm

UnsupportedExtension

Unsupported extension requested in CSR

UnsupportedSignatureAlgorithm

The requested signature algorithm is not supported

RingUnspecified

Unspecified ring error

RingKeyRejected(String)

The ring library rejected the key upon loading

Time

Time conversion related errors

PemError(String)

Error from the pem crate

RemoteKeyError

Error generated by a remote key operation

UnsupportedInCsr

Unsupported field when generating a CSR

InvalidCrlNextUpdate

Invalid certificate revocation list (CRL) next update.

IssuerNotCrlSigner

CRL issuer specifies Key Usages that don’t include cRLSign.

X509(String)

X509 parsing error