Struct rcgen::RevokedCertParams

pub struct RevokedCertParams {
    pub serial_number: SerialNumber,
    pub revocation_time: OffsetDateTime,
    pub reason_code: Option<RevocationReason>,
    pub invalidity_date: Option<OffsetDateTime>,
}

Parameters used for describing a revoked certificate included in a CertificateRevocationList.

Fields
serial_number: SerialNumber

Serial number identifying the revoked certificate.

revocation_time: OffsetDateTime

The date at which the CA processed the revocation.

reason_code: Option<RevocationReason>

An optional reason code identifying why the certificate was revoked.

invalidity_date: Option<OffsetDateTime>

An optional field describing the date on which it was known or suspected that the private key was compromised or the certificate otherwise became invalid. This date may be earlier than the RevokedCertParams::revocation_time.