Trait rcgen::RemoteKeyPair

pub trait RemoteKeyPair {
    fn public_key(&self) -> &[u8];
    fn sign(&self, msg: &[u8]) -> Result<Vec<u8>, RcgenError>;
    fn algorithm(&self) -> &'static SignatureAlgorithm;
}

A private key that is not directly accessible, but can be used to sign messages

Trait objects based on this trait can be passed to the KeyPair::from_remote function to generating certificates from a remote and raw private key, for example an HSM.

Required Methods

fn public_key(&self) -> &[u8]

Returns the public key of this key pair in DER format

fn sign(&self, msg: &[u8]) -> Result<Vec<u8>, RcgenError>

Signs msg using the selected algorithm

fn algorithm(&self) -> &'static SignatureAlgorithm

Reveals which algorithm will be used when you call sign()