Type Alias rcgen::RcgenError

pub type RcgenError = Error;

👎Deprecated: Renamed to Error. We recommend to refer to it by fully-qualifying the crate: rcgen::Error.

Type-alias for the old name of Error.

Aliased Type

enum RcgenError {

    CouldNotParseCertificate,
    CouldNotParseCertificationRequest,
    CouldNotParseKeyPair,
    InvalidIpAddressOctetLength(usize),
    KeyGenerationUnavailable,
    UnsupportedSignatureAlgorithm,
    RingUnspecified,
    RingKeyRejected(String),
    CertificateKeyPairMismatch,
    Time,
    PemError(String),
    RemoteKeyError,
    UnsupportedInCsr,
    InvalidCrlNextUpdate,
    IssuerNotCrlSigner,

}

Variants

CouldNotParseCertificate

The given certificate couldn’t be parsed

CouldNotParseCertificationRequest

The given certificate signing request couldn’t be parsed

CouldNotParseKeyPair

The given key pair couldn’t be parsed

InvalidIpAddressOctetLength(usize)

An IP address was provided as a byte array, but the byte array was an invalid length.

KeyGenerationUnavailable

There is no support for generating keys for the given algorithm

UnsupportedSignatureAlgorithm

The requested signature algorithm is not supported

RingUnspecified

Unspecified ring error

RingKeyRejected(String)

The ring library rejected the key upon loading

CertificateKeyPairMismatch

The provided certificate’s signature algorithm is incompatible with the given key pair

Time

Time conversion related errors

PemError(String)

Error from the pem crate

RemoteKeyError

Error generated by a remote key operation

UnsupportedInCsr

Unsupported field when generating a CSR

InvalidCrlNextUpdate

Invalid certificate revocation list (CRL) next update.

IssuerNotCrlSigner

CRL issuer specifies Key Usages that don’t include cRLSign.